refactor: feedDto commentCount 타입 int -> Long 으로 변경

+++ src/main/java/codeit/sb06/otboo/feed/dto/FeedDto.java
@@ -17,7 +17,7 @@ public record FeedDto(
     List<OotdDto> ootds,
     String content,
     long likeCount,
-    int commentCount,
+    Long commentCount,
     boolean likedByMe
 ) {
   public static FeedDto from(Feed feed) {
